fixed critical bugs
