+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "nodecg-discord-utils",
-    "version": "2.0.0",
+    "version": "2.0.1",
     "description": "Discord voice/text overlay, replicants, bot and API",
     "author": "alphaRomeo323",
     "homepage": "https://github.com/alphaRomeo323/nodecg-discord-utils",
